Comentarios y versiones documentos

+++ backend/proyectosESCOM/proyectosESCOM-ejb/src/main/java/com/mycompany/modulodocumental/entity/DocumentVersion.java
@@ -75,7 +75,7 @@ public class DocumentVersion implements Serializable{
         this.description = description;
     }
 
-    public double getVersion() {
+    public int getVersion() {
         return version;
     }
